--- nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/datastore/internal/TerraformContentDirector.java
@@ -1,45 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2022-present Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-package org.sonatype.nexus.plugins.terraform.datastore.internal;
-
-import javax.inject.Named;
-import javax.inject.Singleton;
-
-import org.sonatype.nexus.repository.Repository;
-import org.sonatype.nexus.repository.content.director.ContentDirector;
-import org.sonatype.nexus.repository.content.fluent.FluentComponent;
-
-/**
- * @since 0.0.6
- */
-@Named("terraform")
-@Singleton
-public class TerraformContentDirector
-        implements ContentDirector
-{
-  // TODO allow move? movement of content? No...
-  @Override
-  public boolean allowMoveTo(final Repository destination) {
-    return true;
-  }
-
-  @Override
-  public boolean allowMoveTo(final FluentComponent component, final Repository destination) {
-    return true;
-  }
-
-  @Override
-  public boolean allowMoveFrom(final Repository source) {
-    return true;
-  }
-}
