+++ nexus-repository-terraform/src/main/java/org/sonatype/nexus/plugins/terraform/internal/attributes/TerraformAttributes.java
@@ -10,21 +10,52 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.plugins.terraform.internal.Attributes;
+package org.sonatype.nexus.plugins.terraform.internal.attributes;
 
 /**
  * @since 0.0.6
  */
-public class TerraformArchiveAttributes extends TerraformVersionAttributes {
+public class TerraformAttributes
+{
+  private String type;
+  private String hostname;
+  private String namespace;
   private String provider;
   private String os;
   private String arch;
+  private String version;
 
+  public String getType() {
+    return type;
+  }
+
+  public TerraformAttributes setType(String type) {
+    this.type = type;
+    return this;
+  }
+
+  public String getHostname() {
+    return hostname;
+  }
+
+  public TerraformAttributes setHostname(String hostname) {
+    this.hostname = hostname;
+    return this;
+  }
+
+  public String getNamespace() {
+    return namespace;
+  }
+
+  public TerraformAttributes setNamespace(String namespace) {
+    this.namespace = namespace;
+    return this;
+  }
   public String getProvider() {
     return provider;
   }
 
-  public TerraformArchiveAttributes setProvider(String provider) {
+  public TerraformAttributes setProvider(String provider) {
     this.provider = provider;
     return this;
   }
@@ -33,7 +64,7 @@ public class TerraformArchiveAttributes extends TerraformVersionAttributes {
     return os;
   }
 
-  public TerraformArchiveAttributes setOs(String os) {
+  public TerraformAttributes setOs(String os) {
     this.os = os;
     return this;
   }
@@ -42,8 +73,17 @@ public class TerraformArchiveAttributes extends TerraformVersionAttributes {
     return arch;
   }
 
-  public TerraformArchiveAttributes setArch(String arch) {
+  public TerraformAttributes setArch(String arch) {
     this.arch = arch;
+    return this;
+  }
+
+  public String getVersion() {
+    return version;
+  }
+
+  public TerraformAttributes setVersion(String version) {
+    this.version = version;
     return this;
   }
